Add checkbox and radio components (#44)
* Add checkbox and radio components

This is an initial first pass at checkbox and radio components. I think
there is still plently of work to be done to make these great, but this
should at least get us started.

* Fix checkbox checked props

* De-duplicate radio/checkbox code as best as possible

This commit remove some of the duplication happening between the radio
and checkbox inputs. It's not perfect, but it at least abstracts the
big parts out. We should continue to clean this up later.