Extrinsically size components
This commit refactors out the concept of "stretched" for most of our
components. This prop started on button (where it is actually useful),
but then ended up being used too much. Most components should simply
fill the width of the container they are giving, thus separating the
concerns of layout and display. This way, the user of can set the size
of the component (and the layout) by simply constraining the width
through whatever layout means they feel work best (e.g. flexbox, fixed
widths, etc). This commit also introduces horizontal checkbox and radio
groups, as they have some specfic layout styling that we want to
preserve, even when allowing the user to determine primary layout.