Fix concatenation of color names when combined with variants (#71)
There are a few places where we try to concatenate the color name with
a variant, but this does not work if the color name you pass is already
a variant. For example: `primaryActive` => `primaryActiveHover` which is
not actually a color name. Instead, we now strip any variants out of the
color name that is passed in a prop.